Rename fingerprint.js to fileSignature.js
Fixes https://github.com/tus/tus-js-client/issues/429
