+++ lib/browser/fileSignature.js
