+++ lib/browser/index.js
@@ -6,7 +6,7 @@ import DetailedError from '../error'
 import { canStoreURLs, WebStorageUrlStorage } from './urlStorage'
 import HttpStack from './httpStack'
 import FileReader from './fileReader'
-import fingerprint from './fingerprint'
+import fingerprint from './fileSignature'
 
 const defaultOptions = {
   ...BaseUpload.defaultOptions,
