+++ lib/node/fileSignature.js
