+++ lib/node/index.js
@@ -6,7 +6,7 @@ import DetailedError from '../error'
 import { FileUrlStorage, canStoreURLs } from './urlStorage'
 import HttpStack from './httpStack'
 import FileReader from './fileReader'
-import fingerprint from './fingerprint'
+import fingerprint from './fileSignature'
 import StreamSource from './sources/StreamSource'
 
 const defaultOptions = {
